backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning and AI
tensorflow==2.15.0
keras==2.15.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
scikit-learn==1.3.2
scipy==1.11.4

# Medical imaging
pydicom==2.4.3
nibabel==5.1.0
SimpleITK==2.3.1

# Image processing
imageio==2.31.6
matplotlib==3.7.2
seaborn==0.12.2

# Database and storage
supabase==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.24.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
loguru==0.7.2
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
